An experimental branch for byte-oriented SAIS.
Turns out to not be a huge difference. The other less-than-ideal
effect is that the length of the suffix array is now the number
of bytes.

My suspicion is that UTF-8 decoding simply isn't a bottleneck.
In particular, we can't take advantage of the smaller alphabet
because of the "w names", which can have very large alphabets.